+++ src/embeds/osu/mod.rs
@@ -140,10 +140,10 @@ pub fn get_map_info(map: &Beatmap, mods: GameMods, stars: f32) -> String {
 
     let attrs = BeatmapAttributesBuilder::default()
         .mode(mode)
-        .ar(map.ar as f64)
-        .cs(map.cs as f64)
-        .hp(map.hp as f64)
-        .od(map.od as f64)
+        .ar(map.ar)
+        .cs(map.cs)
+        .hp(map.hp)
+        .od(map.od)
         .mods(mods.bits())
         .converted(map.convert)
         .build();
